GafferImageUI bindings : Refactor for simplicity
This removes libGafferImageUIBindings, which was no use to anything but the GafferImageUI
module, moving all the code into the module itself.
