+++ src/GafferImageUIModule/ImageViewBinding.cpp
@@ -34,14 +34,93 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#ifndef GAFFERIMAGEUIBINDINGS_IMAGEVIEWBINDING_H
-#define GAFFERIMAGEUIBINDINGS_IMAGEVIEWBINDING_H
+#include "boost/python.hpp"
 
-namespace GafferImageUIBindings
+#include "GafferBindings/NodeBinding.h"
+
+#include "GafferImage/ImageProcessor.h"
+
+#include "GafferImageUI/ImageView.h"
+
+#include "ImageViewBinding.h"
+
+using namespace std;
+using namespace boost::python;
+using namespace Gaffer;
+using namespace GafferBindings;
+using namespace GafferImage;
+using namespace GafferImageUI;
+
+namespace
 {
 
-void bindImageView();
+class ImageViewWrapper : public NodeWrapper<ImageView>
+{
 
-} // namespace GafferImageUIBindings
+	public :
 
-#endif // GAFFERIMAGEUIBINDINGS_IMAGEVIEWBINDING_H
+		ImageViewWrapper( PyObject *self, const std::string &name )
+			:	NodeWrapper<ImageView>( self, name )
+		{
+		}
+
+		void insertConverter( Gaffer::NodePtr converter )
+		{
+			ImageView::insertConverter( converter );
+		}
+
+};
+
+struct DisplayTransformCreator
+{
+	DisplayTransformCreator( object fn )
+		:	m_fn( fn )
+	{
+	}
+
+	ImageProcessorPtr operator()()
+	{
+		IECorePython::ScopedGILLock gilLock;
+		ImageProcessorPtr result = extract<ImageProcessorPtr>( m_fn() );
+		return result;
+	}
+
+	private :
+
+		object m_fn;
+
+};
+
+void registerDisplayTransform( const std::string &name, object creator )
+{
+	ImageView::registerDisplayTransform( name, DisplayTransformCreator( creator ) );
+}
+
+boost::python::list registeredDisplayTransforms()
+{
+	vector<string> n;
+	ImageView::registeredDisplayTransforms( n );
+	boost::python::list result;
+	for( vector<string>::const_iterator it = n.begin(), eIt = n.end(); it != eIt; ++it )
+	{
+		result.append( *it );
+	}
+
+	return result;
+}
+
+} // namespace
+
+void GafferImageUIModule::bindImageView()
+{
+
+	GafferBindings::NodeClass<ImageView, ImageViewWrapper>()
+		.def( init<const std::string &>() )
+		.def( "_insertConverter", &ImageViewWrapper::insertConverter )
+		.def( "registerDisplayTransform", &registerDisplayTransform )
+		.staticmethod( "registerDisplayTransform" )
+		.def( "registeredDisplayTransforms", &registeredDisplayTransforms )
+		.staticmethod( "registeredDisplayTransforms" )
+	;
+
+}
