+++ src/GafferImageUIModule/ImageViewBinding.h
@@ -34,82 +34,14 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#include "boost/python.hpp"
+#ifndef GAFFERIMAGEUIMODULE_IMAGEVIEWBINDING_H
+#define GAFFERIMAGEUIMODULE_IMAGEVIEWBINDING_H
 
-#include "GafferBindings/NodeBinding.h"
-
-#include "GafferImage/ImageProcessor.h"
-
-#include "GafferImageUI/ImageView.h"
-#include "GafferImageUIBindings/ImageViewBinding.h"
-
-using namespace std;
-using namespace boost::python;
-using namespace Gaffer;
-using namespace GafferBindings;
-using namespace GafferImage;
-using namespace GafferImageUI;
-
-class ImageViewWrapper : public NodeWrapper<ImageView>
+namespace GafferImageUIModule
 {
 
-	public :
+void bindImageView();
 
-		ImageViewWrapper( PyObject *self, const std::string &name )
-			:	NodeWrapper<ImageView>( self, name )
-		{
-		}
+} // namespace GafferImageUIModule
 
-};
-
-struct DisplayTransformCreator
-{
-	DisplayTransformCreator( object fn )
-		:	m_fn( fn )
-	{
-	}
-
-	ImageProcessorPtr operator()()
-	{
-		IECorePython::ScopedGILLock gilLock;
-		ImageProcessorPtr result = extract<ImageProcessorPtr>( m_fn() );
-		return result;
-	}
-
-	private :
-
-		object m_fn;
-
-};
-
-static void registerDisplayTransform( const std::string &name, object creator )
-{
-	ImageView::registerDisplayTransform( name, DisplayTransformCreator( creator ) );
-}
-
-static boost::python::list registeredDisplayTransforms()
-{
-	vector<string> n;
-	ImageView::registeredDisplayTransforms( n );
-	boost::python::list result;
-	for( vector<string>::const_iterator it = n.begin(), eIt = n.end(); it != eIt; ++it )
-	{
-		result.append( *it );
-	}
-
-	return result;
-}
-
-void GafferImageUIBindings::bindImageView()
-{
-
-	GafferBindings::NodeClass<ImageView, ImageViewWrapper>()
-		.def( init<const std::string &>() )
-		.def( "_insertConverter", &ImageView::insertConverter )
-		.def( "registerDisplayTransform", &registerDisplayTransform )
-		.staticmethod( "registerDisplayTransform" )
-		.def( "registeredDisplayTransforms", &registeredDisplayTransforms )
-		.staticmethod( "registeredDisplayTransforms" )
-	;
-
-}
+#endif // GAFFERIMAGEUIMODULE_IMAGEVIEWBINDING_H
